% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_CV.R
\name{createCv}
\alias{createCv}
\title{Creates a cross validation column}
\usage{
createCv(data, columName = NULL, kfolds = NULL, seed = NULL)
}
\arguments{
\item{data}{\code{data.frame},  A dataset.}

\item{columName}{\code{vector}, A character vector of the name of one or more columns of the dataset.}

\item{kfolds}{\code{numeric}, Number of folds.}

\item{seed}{\code{numeric}, A seed to specify. By default there is no seed.}
}
\value{
A vector with cross-validation fold index assignment per observation
}
\description{
createCv permit to create a cross validation column by specifiying one or more
columns of the dataset, or a number of folds.
}
\examples{

data(mtcars)

createCv(mtcars, columName = "cyl")
createCv(mtcars, columName = c("cyl", "carb"))
createCv(mtcars, kfolds = 4)
createCv(mtcars, kfolds = 4, seed = 1234)


}
