% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{plot.optiPlusModel}
\alias{plot.optiPlusModel}
\title{Plot}
\usage{
plot.optiPlusModel(x, type, digits = 2, color = "#FF8000")
}
\arguments{
\item{x}{\code{optiPlusModel}.}

\item{type}{\code{character}.}

\item{digits}{\code{numeric}}

\item{color}{\code{character}}
}
\description{
Plot
}
