% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_global_Rf.R
\name{rfMod}
\alias{rfMod}
\title{Launch a RandomForest with cross-validation and the possibility of tuning hyperparameters}
\usage{
rfMod(x, y, cvcol, ntree = 50, mtry = if (!is.null(y) && !is.factor(y))
  max(floor(ncol(x)/3), 1) else floor(sqrt(ncol(x))), maxnodes = NULL,
  nodesize = if (!is.null(y) && !is.factor(y)) 5 else 1, importance = FALSE,
  criterion = "RMSE")
}
\arguments{
\item{x}{\code{data.frame},  Predictor variables.}

\item{y}{\code{vector}, Response variable.}

\item{cvcol}{\code{vector}, Column with cross-validation fold index assignment per observation.}

\item{ntree}{\code{numeric}, Number of trees to grow. Default is 50.}

\item{mtry}{\code{numeric}, Number of variables randomly sampled as candidates at each split. Note that the default values are different for classification (sqrt(p) where p is number of variables in x) and regression (p/3).}

\item{maxnodes}{\code{numeric}, Maximum number of terminal nodes trees in the forest can have. If not given, trees are grown to the maximum possible (subject to limits by nodesize).}

\item{nodesize}{\code{numeric}, Minimum size of terminal nodes. Setting this number larger causes smaller trees to be grown (and thus take less time). Default is 5.}

\item{importance}{\code{numeric}, Should importance of predictors be assessed? Default is FALSE.}

\item{criterion}{\code{character}, Criterion used to select the best model among the grid of hyperparameters.It can be : "RMSE", "R2" or "AUC".}
}
\value{
A list containing :

  \itemize{
    \item{response variable (y)}
    \item{predicted values (yp)}
    \item{the cross-validation column (cvcol)}
    \item{the optimized parameters (ntree, mtry, maxnodes, nodesize)}
    \item{the criteria (RMSE, R2, AUC...)}
  }
}
\description{
rfMod uses the {randomForest} function of the package {randomForest}. rfMod allows to launch a random forest
for classification and regression while choosing a column of cross-validation and specifying a grid of hyperparameters.
}
\examples{

data(mtcars)

#Creation of cross-validation column :
set.seed(1234)
cv <- sample(1:8, nrow(mtcars), replace = TRUE)

#Data
y <- "mpg"
ycolumnindex <- names(mtcars) == "mpg"
x <- mtcars[, !ycolumnindex]
y <- mtcars[, ycolumnindex]

rfMod(x = x, y = y, cvcol= cv,
 ntree= c(50, 100), mtry = c(3,4),
  nodesize = c(3, 4, 5),  criterion = "RMSE")


}
